Add note explaining error bars
The alignment here isn't perfect, but I think there's basically no way
to get this exactly right within Plotly. I did attempt doing this in
HTML and positioning it over the plot but then it wouldn't show up in
the PNG download and so I think this is overall better.

Closes #87