Fix bug with CCG/Lab list sizes
The list size for a CCG or Lab should always be the sum of the list
sizes of its practices. If we are filtering by a certain test code and
one practice has no results at all for that test code its list size
should still be included in the CCG list size.