Persist highlight_entities when changing groupby
This is an attempt to address #71. Storing the highlighted entities in
cookies feels like a bad move to me because I think it will lead to
unexpected edge cases (like things being inadvertantly hightlighted in
measure charts) which other users then can't reproduce.

It's true that the highlighted entities become no longer relevant when
changing the group by, but they won't do any harm either and they'll be
ready and waiting when you switch back to the original group by.

If we had more time we could refactor things so that the entity IDs in
the query string were prefix by their type. This would mean there
wouldn't be any possiblility of accidental matches. As it stands though,
the ID space of our types is almost entirely disjoint. (There's a bit of
overlap with some error codes and low pratice IDs, but this will
disappear if we stop annonymising practices.)

Closes #71