Fix list size issues with labs and CCGs
This PR attempts to address the two issues raised in #140 which result
in odd list sizes when grouping by CCG or lab.

The CCG issue is caused by the fact that we were extracting list sizes
from our test count data which meant that where a practice ordered no
tests at all in a given month its list size would disappear from the CCG
total for that month. The solution is to get the list sizes directly
from the practice_codes file.

However we can't do this straightforwardly as things stand because the
practice code anonymisation means that we can't match up this file with
our test count data. So instead we calculate the correct total for each
CCG in each month and insert dummy entries into our list size dataframe
so that the totals come out correct when grouping by CCG. As noted in
the comments, once we stop anonymising we can just drop this code.

The lab issue is that there is no straightforward definition of list
size for labs. The most naive option to say that a lab's list size is
the sum of the list sizes of all practices which ordered any tests from
it during the month. As noted, this causes lab list sizes to fluctuate
in a way which renders them not particularly useful as a denominator.

The solution we use in this PR is to assign each practice to the lab
where it sent the majority of its potassium tests, ignoring practices
which sent less than 50 potassium tests to any one lab. This implies
that some practices will not have any associated lab in some months, but
it at least keeps the lab list sizes reasonably stable.

Closes #140