Add a multiple selector input for selecting organisations
This provides an alternative entity-choosing interface to the
heatmap.  This is required for someone who wishes to browse the data
starting from a practice.

Part of this addition involves changing how the heatmap and
querystring are linked; we now go via page_stage, which is actually
more consistent with the way we handle the path part of the URL.